[KYUUBI #3934] Merge the class to TrinoContext
