+++ kyuubi-server/src/main/scala/org/apache/kyuubi/server/trino/api/TrinoContext.scala
@@ -18,34 +18,38 @@
 package org.apache.kyuubi.server.trino.api
 
 import java.io.UnsupportedEncodingException
-import java.net.{URLDecoder, URLEncoder}
+import java.net.{URI, URLDecoder, URLEncoder}
+import java.util
 import javax.ws.rs.core.{HttpHeaders, Response}
 
 import scala.collection.JavaConverters._
 
+import io.trino.client.{ClientTypeSignature, Column, QueryError, QueryResults, StatementStats, Warning}
 import io.trino.client.ProtocolHeaders.TRINO_HEADERS
-import io.trino.client.QueryResults
+import org.apache.hive.service.rpc.thrift.{TGetResultSetMetadataResp, TRowSet}
+
+import org.apache.kyuubi.operation.OperationStatus
 
 /**
  * The description and functionality of trino request
  * and response's context
  *
- * @param user Specifies the session user, must be supplied with every query
- * @param timeZone  The timezone for query processing
+ * @param user               Specifies the session user, must be supplied with every query
+ * @param timeZone           The timezone for query processing
  * @param clientCapabilities Exclusive for trino server
- * @param source This supplies the name of the software that submitted the query,
- *               e.g. `trino-jdbc` or `trino-cli` by default
- * @param catalog The catalog context for query processing, will be set response
- * @param schema The schema context for query processing
- * @param language The language to use when processing the query and formatting results,
- *               formatted as a Java Locale string, e.g., en-US for US English
- * @param traceToken Trace token for correlating requests across systems
- * @param clientInfo Extra information about the client
- * @param clientTags Client tags for selecting resource groups. Example: abc,xyz
- * @param preparedStatement `preparedStatement` are kv pairs, where the names
- *                          are names of previously prepared SQL statements,
- *                          and the values are keys that identify the
- *                          executable form of the named prepared statements
+ * @param source             This supplies the name of the software that submitted the query,
+ *                           e.g. `trino-jdbc` or `trino-cli` by default
+ * @param catalog            The catalog context for query processing, will be set response
+ * @param schema             The schema context for query processing
+ * @param language           The language to use when processing the query and formatting results,
+ *                           formatted as a Java Locale string, e.g., en-US for US English
+ * @param traceToken         Trace token for correlating requests across systems
+ * @param clientInfo         Extra information about the client
+ * @param clientTags         Client tags for selecting resource groups. Example: abc,xyz
+ * @param preparedStatement  `preparedStatement` are kv pairs, where the names
+ *                           are names of previously prepared SQL statements,
+ *                           and the values are keys that identify the
+ *                           executable form of the named prepared statements
  */
 case class TrinoContext(
     user: String,
@@ -62,6 +66,11 @@ case class TrinoContext(
     preparedStatement: Map[String, String] = Map.empty) {}
 
 object TrinoContext {
+
+  private val defaultWarning: util.List[Warning] = new util.ArrayList[Warning]()
+  private val GENERIC_INTERNAL_ERROR_CODE = 65536
+  private val GENERIC_INTERNAL_ERROR_NAME = "GENERIC_INTERNAL_ERROR_NAME"
+  private val GENERIC_INTERNAL_ERROR_TYPE = "INTERNAL_ERROR"
 
   def apply(headers: HttpHeaders): TrinoContext = {
     apply(headers.getRequestHeaders.asScala.toMap.map {
@@ -165,5 +174,220 @@ object TrinoContext {
       case e: UnsupportedEncodingException =>
         throw new AssertionError(e)
     }
+
+  def createQueryResults(
+      queryId: String,
+      nextUri: URI,
+      queryHtmlUri: URI,
+      queryStatus: OperationStatus,
+      columns: Option[TGetResultSetMetadataResp] = None,
+      data: Option[TRowSet] = None): QueryResults = {
+
+    //    val queryHtmlUri = uriInfo.getRequestUriBuilder
+    //      .replacePath("ui/query.html").replaceQuery(queryId).build()
+
+    val columnList = columns match {
+      case Some(value) => convertTColumn(value)
+      case None => null
+    }
+    val rowList = data match {
+      case Some(value) => convertTRowSet(value)
+      case None => null
+    }
+
+    new QueryResults(
+      queryId,
+      queryHtmlUri,
+      nextUri,
+      nextUri,
+      columnList,
+      rowList,
+      StatementStats.builder.setState(queryStatus.state.name()).build(),
+      toQueryError(queryStatus),
+      defaultWarning,
+      null,
+      0L)
+  }
+
+  def convertTColumn(columns: TGetResultSetMetadataResp): util.List[Column] = {
+    columns.getSchema.getColumns.asScala.map(c => {
+      val tp = c.getTypeDesc.getTypes.get(0).getPrimitiveEntry.getType.name()
+      new Column(c.getColumnName, tp, new ClientTypeSignature(tp))
+    }).toList.asJava
+  }
+
+  def convertTRowSet(rowSet: TRowSet): util.List[util.List[Object]] = {
+    var dataSet: Array[scala.List[Object]] = Array()
+
+    if (rowSet.getColumns == null) {
+      return rowSet.getRows.asScala
+        .map(t => t.getColVals.asScala.map(v => v.getFieldValue.asInstanceOf[Object]).asJava)
+        .asJava
+    }
+
+    rowSet.getColumns.asScala.foreach {
+      case tColumn if tColumn.isSetBoolVal =>
+        val nulls = util.BitSet.valueOf(tColumn.getBoolVal.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getBoolVal.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getBoolVal.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetByteVal =>
+        val nulls = util.BitSet.valueOf(tColumn.getByteVal.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getByteVal.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(scala.List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(scala.List(x._1))
+            }
+        } else {
+          tColumn.getByteVal.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetI16Val =>
+        val nulls = util.BitSet.valueOf(tColumn.getI16Val.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getI16Val.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getI16Val.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetI32Val =>
+        val nulls = util.BitSet.valueOf(tColumn.getI32Val.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getI32Val.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getI32Val.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetI64Val =>
+        val nulls = util.BitSet.valueOf(tColumn.getI64Val.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getI64Val.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getI64Val.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetDoubleVal =>
+        val nulls = util.BitSet.valueOf(tColumn.getDoubleVal.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getDoubleVal.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getDoubleVal.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn if tColumn.isSetBinaryVal =>
+        val nulls = util.BitSet.valueOf(tColumn.getBinaryVal.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getBinaryVal.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getBinaryVal.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+      case tColumn =>
+        val nulls = util.BitSet.valueOf(tColumn.getStringVal.getNulls)
+        if (dataSet.isEmpty) {
+          dataSet = tColumn.getStringVal.getValues.asScala.zipWithIndex
+            .foldLeft(Array[scala.List[Object]]()) {
+              case (acc, x) if nulls.get(x._2) =>
+                acc ++ List(List(None))
+              case (acc, x) if !nulls.get(x._2) =>
+                acc ++ List(List(x._1))
+            }
+        } else {
+          tColumn.getStringVal.getValues.asScala.zipWithIndex.foreach {
+            case (_, rowIdx) if nulls.get(rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(None)
+            case (v, rowIdx) =>
+              dataSet(rowIdx) = dataSet(rowIdx) ++ List(v)
+          }
+        }
+    }
+    dataSet.toList.map(_.asJava).asJava
+  }
+
+  def toQueryError(queryStatus: OperationStatus): QueryError = {
+    val exception = queryStatus.exception
+    if (exception.isEmpty) {
+      null
+    } else {
+      new QueryError(
+        exception.get.getMessage,
+        queryStatus.state.name(),
+        GENERIC_INTERNAL_ERROR_CODE,
+        GENERIC_INTERNAL_ERROR_NAME,
+        GENERIC_INTERNAL_ERROR_TYPE,
+        null,
+        null)
+    }
+  }
 
 }
